card_recommendation/requirements.txt
Django>=5.1.7
django-heroku
gunicorn
psycopg2-binary
djangorestframework>=3.14.0
djangorestframework-jsonapi>=7.1.0
djangorestframework-simplejwt
python-dotenv
dj-database-url
whitenoise
twilio
xai-sdk>=1.3.1
PyPDF2>=3.0.0
